refactor(ui): reusable QR transfer control (no navigation for onboarding restore)
The QR backup/restore UI was a full page everyone navigated to — which dropped
onboarding users onto a framed, logged-in-looking screen. Extract it into a
reusable <QRTransfer> and render it in place.

- New components/QRTransfer.tsx holds the whole generate/scan/restore flow
  (camera + jsQR logic unchanged), taking mode / initialTab / onRestored props.
- /backup/qr becomes a thin wrapper: RouteGuard + AppLayout + <QRTransfer>
  (keeps the ?tab=restore deep link + restore-tab wallet-gate exemption).
- BackupQRModal now renders <QRTransfer> inline in a dialog (desktop) / drawer
  (mobile) instead of navigating away; it mounts only while open so its unmount
  cleanup stops the camera. Onboarding restore stays on /onboarding.

e2e: onboarding "Scan QR codes" opens the QR UI inline without navigating.
typecheck · build · 35 e2e.
